% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findn_3pod.R
\name{findn_3pod}
\alias{findn_3pod}
\title{Find the Sample Size Using the 3pod Algorithm}
\usage{
findn_3pod(fun, targ, start, k = 50, init_evals = 100, max_evals = 1000, ...)
}
\arguments{
\item{fun}{A function that estimates the power of a trial. The function has
to take at least two arguments: n, the sample size and k, the number of
iterations.}

\item{targ}{The target power.}

\item{start}{An initial guess for the sample size.}

\item{k}{How often the target function is evaluated at each design point.}

\item{init_evals}{How many evaluations the first model is based on.}

\item{max_evals}{The maximum number of simulations.}

\item{...}{Further optional arguments.}
}
\description{
\code{findn_3pod} estimates the sample size for a certain target function 
based on a modified version of the 3pod algorithm (Wu & Tian, 2014).
}
