% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findn_3pod.R
\name{findn_3pod}
\alias{findn_3pod}
\title{Find the Sample Size Using the 3pod Algorithm}
\usage{
findn_3pod(fun, targ, start, k = 50, startevals = 100, maxevals = 1000, ...)
}
\arguments{
\item{fun}{A target function that estimates the power of a trial.}

\item{targ}{The target power. Must bei either 0.8 or 0.9.}

\item{start}{An initial guess for the sample size.}

\item{k}{How often the target function is evaluated at each design point.}

\item{startevals}{How many evaluations are used for the first phase.}

\item{maxevals}{Maximum number of function evaluations.}

\item{...}{Further arguments to be passed to \code{findn_3pod}.}
}
\description{
\code{findn_3pod} estimates the sample size for a certain target function 
based on a modified version of the 3pod algorithm (Wu & Tian, 2014).
}
