% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findn_maruo.R
\name{findn_maruo}
\alias{findn_maruo}
\title{Find the Sample Size Using the Algorithm by Maruo et al.}
\usage{
findn_maruo(fun, targ, start = 10, k = 100, ...)
}
\arguments{
\item{fun}{A function that estimates the power of a trial. The function has
to take at least two arguments: n, the sample size and k, the number of
iterations.}

\item{targ}{The target power. Must be either 0.8 or 0.9.}

\item{start}{Starting value for the algorithm. Maruo et al. suggest to
use 10.}

\item{k}{Number of trial simulations to use in \code{fun} to estimate the power.}

\item{...}{Further optional arguments.}
}
\description{
\code{findn_maruo} estimates the sample size for a certain target function 
based on repeated simulations using a model based approach proposed by Maruo 
et al. (2018).
}
\references{
Maruo, K., Tada, K., Ishil, R. and Gosho M. (2018) An
  Efficient Procedure for Calculating Sample Size Through
  Statistical Simulations, Statistics in Biopharmaceutical Research
  10, 1-8.  \doi{https://doi.org/10.1080/19466315.2017.1349689}
}
