% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findn_maruo.R
\name{findn_maruo}
\alias{findn_maruo}
\title{Find the Sample Size Using the Algorithm by Maruo et al.}
\usage{
findn_maruo(fun, targ, start = 10, k = 100, ...)
}
\arguments{
\item{fun}{A function that estimates the power of a trial. The function has
to take at least two arguments: n, the sample size and k, the number of
iterations.}

\item{start}{Starting value for the algorithm. Maruo et al. suggest to
use 10.}

\item{k}{How often the target function is evaluated at each design point.}

\item{...}{Further optional arguments.}

\item{targ.}{The target power. Must be either 0.8 or 0.9.}
}
\description{
\code{findn_maruo} estimates the sample size for a certain target function 
based on repeated simulations using a model based approach proposed by Maruo 
et al. (2018).
}
