% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findn_rm.R
\name{findn_rm}
\alias{findn_rm}
\title{Find the Sample Size Using the Robbing-Monro Algorithm}
\usage{
findn_rm(fun, targ, start, k = 50, maxevals = 1000, j = 1, avg = 0.5, ...)
}
\arguments{
\item{fun}{A target function that estimates the power of a trial.}

\item{targ}{The target power. Must bei either 0.8 or 0.9.}

\item{start}{An initial guess for the sample size.}

\item{k}{How often the target function is evaluated at each design point.}

\item{maxevals}{Maximum number of function evaluations.}

\item{j}{A tuning parameter for choosing the step size. In the original
Robbins-Monro algorithm j is 1.}

\item{avg}{A tuning parameter for the averaged veresion of the Robbins-Monro
algorithm. The sample size estimate is the average of the last avg% of
design points in the algorithm.}

\item{...}{Further arguments to be passed to \code{findn_rm}.}
}
\description{
\code{findn_rm} estimates the sample size for a certain target function
using a modified version of the Robbins-Monro algorithm.
}
